Update electron-log and fix renderer logging
