+++ planetwars/client/app/process/index.ts
@@ -1,0 +1,7 @@
+import * as main from './main';
+import * as shared from './shared';
+
+export {
+  main,
+  shared,
+};
