+++ planetwars/client/app/process/main/index.ts
@@ -1,4 +1,3 @@
-export * from './logging';
 export * from './app_listeners';
 export * from './window';
 export * from './menu';
