+++ planetwars/client/package.json
@@ -114,7 +114,7 @@
     "d3": "^5.5.0",
     "electron-context-menu": "^0.11.0",
     "electron-debug": "^2.0.0",
-    "electron-log": "^2.2.17",
+    "electron-log": "^3.0.1",
     "electron-updater": "4.0.6",
     "electron-window-state": "^5.0.3",
     "font-awesome": "^4.7.0",
